prometheus: Add not quite trivial encoding tests
Using parameterization, check a small sample of folder configurations.
These are already tested in the normal `Backlog` tests (and more
extensively), but here we're checking the presence of various elements
of the metrics.

The coverage metrics are likely somewhat inflated, since all encodings
are run, but not all are verified, but well, the most important are, I
think.