Fix for gumdrop's behaviour when not using `_exit`
When not using the `_exit` variants, gumdrop doesn't show the usage
at all, relying on the caller to handle this. This is suboptimal, since
there's no full "show usage" functionality - the `usage()` only shows
parts of it. So this is a quick workaround, need to think more on how
to a) handle this, b) prevent regressions in the future.