Fix completely missing dir/file mode passing
Sigh. I relied on tests, but no tests were checking actual
CLI-to-collector passing, as most tests were directly exercising only
the backlog scanner, not the end-to-end integration. Fix this for real
by adding tests for the oneshot binary (since that's easy to test, being
a simple CLI tool), and the `assert_cmd` crate, which is pretty nice but
pulls in quite a bit of (dev) dependencies. Well, you can't have it all.